+++ gradle.properties
@@ -3,7 +3,7 @@
     org.gradle.parallel=true
     org.gradle.caching=true
 # Common library metadata
-    version=0.1.5
+    version=0.1.2
     group=com.enderzombi102.enderlib
     repo=https://github.com/ENDERZOMBI102/EnderLib
     archivesBaseName=enderlib
@@ -11,4 +11,4 @@
     annotations_version=23.0.0
     slf4j_version=1.8-beta4
     log4j_version=2.17.2
-    manifold_version=2022.1.19
+    manifold_version=2022.1.10
