--- minecraft/gradle.properties
@@ -1,10 +1,0 @@
-# Loader Properties
-    minecraft_version=1.19.2
-    quilt_mappings=8
-    loader_version=0.17.4
-# Mod Metadata
-    archivesBaseName=enderlib-minecraft
-# Dependencies
-    qfapi_version=4.0.0-beta.9+0.60.0
-    modmenu_version=4.0.6
-
