--- minecraft/src/main/resources/enderlib.mixins.json
@@ -1,14 +1,0 @@
-{
-	"required": true,
-	"minVersion": "0.8",
-	"package": "com.enderzombi102.enderlib.minecraft.impl.mixin",
-	"compatibilityLevel": "JAVA_17",
-	"mixins": [
-		"WorldMixin",
-		"LanguageMixin"
-	],
-	"client": [ ],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
