--- minecraft/src/main/resources/quilt.mod.json
@@ -1,45 +1,0 @@
-{
-	"schema_version": 1,
-	"quilt_loader": {
-		"group": "com.enderzombi102.enderlib.minecraft",
-		"id": "enderlib-minecraft",
-		"version": "${version}",
-		"metadata": {
-			"name": "EnderLib Minecraft",
-			"description": "${description} QuiltMC ( jvm-17 )",
-			"contributors": {
-				"ENDERZOMBI102": "Owner"
-			},
-			"contact": {
-				"issues": "${repo}/issues",
-				"sources": "${repo}"
-			},
-			"icon": "assets/enderlib/enderlib-quilt.png"
-		},
-		"intermediate_mappings": "net.fabricmc:intermediary",
-		"entrypoints": {
-			"modmenu": "com.enderzombi102.enderlib.minecraft.impl.compat.modmenu.ModMenuCompat"
-		},
-		"depends": [
-			{
-				"id": "quilt_loader",
-				"versions": ">=0.16.0"
-			},
-			{
-				"id": "minecraft",
-				"versions": ">=1.18.2"
-			}
-		]
-	},
-	"mixin": "enderlib.mixins.json",
-	"modmenu": {
-		"badges": [ "library" ],
-		"parent": {
-			"id": "enderlib",
-			"name": "EnderLib",
-			"description": "A set of libraries to ease ENDERZOMBI102's java experience.",
-			"icon": "assets/enderlib/enderlib.png",
-			"badges": [ "library" ]
-		}
-	}
-}
