--- minecraft/src/testmod/resources/assets/enderlib-testmod/blockstates/testblock.json
@@ -1,7 +1,0 @@
-{
-	"variants": {
-		"": {
-			"model": "enderlib-testmod:block/testblock"
-		}
-	}
-}
