--- minecraft/src/testmod/resources/assets/enderlib-testmod/lang/en_us.json
@@ -1,5 +1,0 @@
-{
-	"block.enderlib-testmod": {
-		"testblock": "Test Block"
-	}
-}
