--- minecraft/src/testmod/resources/assets/enderlib-testmod/models/block/testblock.json
@@ -1,6 +1,0 @@
-{
-	"parent": "block/cube_all",
-	"textures": {
-		"all": "enderlib-testmod:block/testblock"
-	}
-}
