--- minecraft/src/testmod/resources/assets/enderlib-testmod/models/item/testblock.json
@@ -1,3 +1,0 @@
-{
-	"parent": "enderlib-testmod:block/testblock"
-}
