--- minecraft/src/testmod/resources/testmod.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"minVersion": "0.8",
-	"package": "com.enderzombi102.enderlib.test.mixin",
-	"compatibilityLevel": "JAVA_17",
-	"mixins": [ ],
-	"client": [ ],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
