+++ settings.gradle
@@ -1,13 +1,13 @@
 pluginManagement {
 	repositories {
-		gradlePluginPortal()
 		maven { url 'https://maven.fabricmc.net/' }
 		maven { url 'https://maven.quiltmc.org/repository/release' }
-		maven { url 'https://maven.quiltmc.org/repository/snapshot' }
+		gradlePluginPortal()
 	}
 }
 
 include 'jvm8'
-include 'minecraft'
+include 'quilt'
+include 'fabric'
 include 'manifolds'
 include 'manifolds-jvm8'
