--- src/test/java/Private.java
@@ -1,11 +1,0 @@
-public class Private {
-	private static int zero = 0;
-
-	private static void increment() {
-		zero += 1;
-	}
-
-	public static int getZero() {
-		return zero;
-	}
-}
