--- src/test/java/Private1.java
@@ -1,1 +1,0 @@
-public record Private1( int num ) { }
