--- src/test/java/Valve.java
@@ -1,7 +1,0 @@
-public enum Valve {
-	One,
-	Two,
-	TwoEp1,
-	TwoEp2,
-	Alyx
-}
