--- todo.yml
@@ -1,2 +1,0 @@
-- minecraft:
-   - TestMod should test BlockBreakEvent
